Small tweak in forgeBSgenomeDataPkgFromNCBI()'s internal sanity checks
Depending on whether a GenBank or RefSeq assembly accession is provided,
the sequences in the assembly report that have an NA in the GenBankAccn
or RefSeqAccn column should not be expected to be found in the
corresponding FASTA file. So the various internal sanity checks performed
by forgeBSgenomeDataPkgFromNCBI() now ignore them.